Updated jstree to latest from Github pre1.0 branch, #88
Updating to the latest code from Github fixed IE not reading the
tree hierarchy correctly. Still buggy in IE but this helps, and
jstree 1.0 is close too, so upgrading is likely the best path
forward once that's out. None of the other libraries handle
trees, so replacement with another library would be much more
difficult.