Fixed Controller setting the wrong app and params for some nested URLs
For URLs that point to pages and not handlers, but contain additional
URL elements after the page ID (e.g., `/some-page/two/three`),
`Controller::route()` was setting `$this->app` to the first element,
and not adding it to `$this->params`. This causes the default handler
(`admin/page`) to use the second URL element as the page ID instead
of the first.